requirements.txt
apify_client==2.5.0
fastapi==0.136.0
openpyxl==3.1.5
pydantic==2.13.2
python-dotenv==1.2.2
uvicorn==0.44.0
